moved product_elements
